more un-templating and file splitting
