+++ CMakeLists.txt
@@ -200,12 +200,14 @@ set(objs
     ${CMAKE_SOURCE_DIR}/src/functionparser.cpp
     ${CMAKE_SOURCE_DIR}/src/random.cpp
     ${CMAKE_SOURCE_DIR}/src/mpi.cpp
+    ${CMAKE_SOURCE_DIR}/src/chainmove.cpp
+    ${CMAKE_SOURCE_DIR}/src/clustermove.cpp
     ${CMAKE_SOURCE_DIR}/src/core.cpp
     ${CMAKE_SOURCE_DIR}/src/units.cpp
     ${CMAKE_SOURCE_DIR}/src/analysis.cpp
     ${CMAKE_SOURCE_DIR}/src/atomdata.cpp
     ${CMAKE_SOURCE_DIR}/src/energy.cpp
-    ${CMAKE_SOURCE_DIR}/src/external_potential.cpp
+    ${CMAKE_SOURCE_DIR}/src/externalpotential.cpp
     ${CMAKE_SOURCE_DIR}/src/geometry.cpp
     ${CMAKE_SOURCE_DIR}/src/group.cpp
     ${CMAKE_SOURCE_DIR}/src/io.cpp
@@ -226,9 +228,11 @@ set(hdrs
     ${CMAKE_SOURCE_DIR}/src/average.h
     ${CMAKE_SOURCE_DIR}/src/atomdata.h
     ${CMAKE_SOURCE_DIR}/src/auxiliary.h
+    ${CMAKE_SOURCE_DIR}/src/chainmove.h
+    ${CMAKE_SOURCE_DIR}/src/clustermove.h
     ${CMAKE_SOURCE_DIR}/src/core.h
     ${CMAKE_SOURCE_DIR}/src/energy.h
-    ${CMAKE_SOURCE_DIR}/src/external_potential.h
+    ${CMAKE_SOURCE_DIR}/src/externalpotential.h
     ${CMAKE_SOURCE_DIR}/src/functionparser.h
     ${CMAKE_SOURCE_DIR}/src/geometry.h
     ${CMAKE_SOURCE_DIR}/src/group.h
