+++ src/externalpotential.cpp
@@ -1,4 +1,4 @@
-#include "external_potential.h"
+#include "externalpotential.h"
 
 namespace Faunus {
 namespace Energy {
