+++ src/externalpotential.h
